Rename socket proxy helper to bridge
