+++ src-tauri/src/lib.rs
@@ -9,16 +9,16 @@ pub mod overlay;
 pub mod runtime;
 pub mod secrets;
 pub mod server;
-pub mod socket_proxy;
+pub mod socket_bridge;
 pub mod status;
 pub mod tools_cache;
 pub mod transport;
 pub mod types;
 
-pub const BIN_NAME_SOCKET_PROXY: &str = if cfg!(windows) {
-    "mcp-bouncer-socket-proxy.exe"
+pub const BIN_NAME_SOCKET_BRIDGE: &str = if cfg!(windows) {
+    "mcp-bouncer-socket-bridge.exe"
 } else {
-    "mcp-bouncer-socket-proxy"
+    "mcp-bouncer-socket-bridge"
 };
 
 pub mod logging {
