+++ src-tauri/src/socket_bridge.rs
